Implemented transaction controllers. Minor fixes

+++ Application/Interfaces/Services/ICategoryService.cs
@@ -6,6 +6,6 @@ public interface ICategoryService
 {
     public CategoryDto AddCategory(CategoryDto categoryDto);
     public CategoryDto UpdateCategory(CategoryDto categoryDto);
-    public CategoryDto DeleteCategory(int id);
+    public CategoryDto DeleteCategory(long id);
     public List<CategoryDto> GetCategories(CategoryFilterDto filter);
 }
